requirements.txt
streamlit==1.40.0

# LangChain 1.x family (compatible with langchain-groq 1.1.1)
langchain==1.1.3
langchain-core==1.2.0
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-groq==1.1.1

# Vector search & retrievers
faiss-cpu==1.13.1
rank-bm25==0.2.2

# Embeddings & rerankers
transformers==4.46.2
sentence-transformers==3.2.1
torch==2.2.2

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
tqdm==4.66.5
pydantic==2.9.2

# PDF reading
pypdf==5.1.0

datasets==2.20.0
